+++ CHANGELOG.md
@@ -7,6 +7,7 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## Added
 - **Experimental**, gateway: add UUID support for VPN gateway connections and tunnels
+- service: a more elaborate error message if get request returns an error on unmarshalling json array
 
 ## [8.5.0]